+++ src/components/feed/feed-order/ProfileOrder.tsx
@@ -1,44 +1,15 @@
-import { useEffect } from 'react';
-import styles from './ProfileOrder.module.css';
 import { useTypedSelector } from '../../../hooks/useTypeSelector';
+import styles from './ProfileOrder.module.css';
 import { ListFeedOrders } from './list-feed-order/ListFeedOrders';
-import { useAppDispatch } from '../../../hooks/hooks';
 import { TWSState } from '../../../services/reducers/ws';
 import { getWSSelector } from '../../../services/selectors/selector';
-import { WSActionTypes } from '../../../services/store/types/ws';
-import { getCookie } from '../../../services/common/common';
-import { NORMA_API_ORDERS_WS } from '../../../data/data';
 
 function ProfileOrder() {
-  const dispatch = useAppDispatch();
-
   // Получаем данные из хранилища redux.
   const { error, messages, wsConnected } =
     useTypedSelector<TWSState>(getWSSelector);
 
   const { orders } = messages;
-
-  useEffect(() => {
-    // Открытие wev socket.
-    const accessToken = getCookie('accessToken');
-    let token = '';
-    if (accessToken) {
-      token = accessToken.split(' ')[1];
-    }
-
-    dispatch({
-      type: WSActionTypes.WS_CONNECTION_START,
-      payload: NORMA_API_ORDERS_WS + '?token=' + token,
-    });
-
-    return () => {
-      // Закрытие web socket.
-      dispatch({
-        type: WSActionTypes.WS_CONNECTION_CLOSED,
-        payload: '',
-      });
-    };
-  }, []);
 
   return !wsConnected ? (
     <section className={styles['Info-container']}>
