+++ src/hooks/hooks.ts
@@ -1,5 +1,4 @@
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { AppDispatch, RootState } from '../services/store/store';
+import { useDispatch } from 'react-redux';
+import type { AppDispatch } from '../services/store/store';
 
 export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
